moved reaction xp gain to levels cog
